+++ airflow_e2e/composer/composer.py
@@ -1,13 +1,13 @@
 from pathlib import Path
 
-from airflow_e2e.builder.builder import build_airflow_core_services_composer
+from airflow_e2e.builder.builder import (
+    build_airflow_core_services_composer,
+    build_e2e_test_runner_service_composer,
+)
 from airflow_e2e.composer.airflow_seeder_service_composer import (
     AirflowSeederServiceComposer,
 )
 from airflow_e2e.composer.constants import DOCKER_FOLDER_NAME
-from airflow_e2e.composer.e2e_test_runner_service_composer import (
-    E2eTestRunnerServiceComposer,
-)
 from airflow_e2e.composer.envrc_file_writer import EnvrcFileWriter
 from airflow_e2e.composer.manual_e2e_test_runner_service_composer import (
     ManualE2eTestRunnerServiceComposer,
@@ -19,6 +19,7 @@ def setup(
     tests: str,
     working_dir: str,
     with_custom_airflow_packages: bool,
+    with_custom_test_packages: bool,
 ):
     docker_folder_path = Path(working_dir) / DOCKER_FOLDER_NAME
     docker_folder_path.mkdir(parents=True, exist_ok=True)
@@ -29,8 +30,10 @@ def setup(
     )
     airflow_core_services_composer.setup(working_dir=docker_folder_path)
 
-    e2e_test_runner_service_composer = E2eTestRunnerServiceComposer(
-        dags=dags, tests=tests
+    e2e_test_runner_service_composer = build_e2e_test_runner_service_composer(
+        dags=dags,
+        tests=tests,
+        with_custom_test_packages=with_custom_test_packages,
     )
     e2e_test_runner_service_composer.setup(working_dir=docker_folder_path)