+++ tests/airflow_e2e/composer/test_composer.py
@@ -14,6 +14,7 @@ def test_should_create_docker_base_folder():
             tests="some/tests/folder",
             working_dir=temp_dir,
             with_custom_airflow_packages=False,
+            with_custom_test_packages=False,
         )
 
         expected_docker_folder_path = Path(temp_dir) / "docker"
@@ -37,6 +38,7 @@ def test_setup_should_setup_airflow_core_services(
             tests="some/tests/folder",
             working_dir=temp_dir,
             with_custom_airflow_packages=with_custom_airflow_packages_flag,
+            with_custom_test_packages=False,
         )
 
         assert spy_build.call_count == 1
@@ -51,10 +53,13 @@ def test_setup_should_setup_airflow_core_services(
         )
 
 
-def test_setup_should_setup_e2e_test_runner_service(mocker):
+@pytest.mark.parametrize("with_custom_test_packages_flag", [True, False])
+def test_setup_should_setup_e2e_test_runner_service(
+    mocker, with_custom_test_packages_flag: bool
+):
     mock_composer_instance = MagicMock()
-    spy_composer = mocker.patch(
-        "airflow_e2e.composer.composer.E2eTestRunnerServiceComposer",
+    spy_build = mocker.patch(
+        "airflow_e2e.composer.composer.build_e2e_test_runner_service_composer",
         return_value=mock_composer_instance,
     )
 
@@ -64,12 +69,14 @@ def test_setup_should_setup_e2e_test_runner_service(mocker):
             tests="some/tests/folder",
             working_dir=temp_dir,
             with_custom_airflow_packages=False,
+            with_custom_test_packages=with_custom_test_packages_flag,
         )
 
-        assert spy_composer.call_count == 1
-        assert spy_composer.call_args == call(
+        assert spy_build.call_count == 1
+        assert spy_build.call_args == call(
             dags="some/dags/folder",
             tests="some/tests/folder",
+            with_custom_test_packages=with_custom_test_packages_flag,
         )
 
         assert mock_composer_instance.setup.call_count == 1
@@ -91,6 +98,7 @@ def test_setup_should_setup_airflow_seeder_service_composer(mocker):
             tests="some/tests/folder",
             working_dir=temp_dir,
             with_custom_airflow_packages=False,
+            with_custom_test_packages=False,
         )
 
         assert spy_composer.call_count == 1
@@ -115,6 +123,7 @@ def test_setup_should_setup_manual_e2e_test_runner_service_composer(mocker):
             tests="some/tests/folder",
             working_dir=temp_dir,
             with_custom_airflow_packages=False,
+            with_custom_test_packages=False,
         )
 
         assert spy_composer.call_count == 1
@@ -139,6 +148,7 @@ def test_setup_should_setup_envrc_file_writer(mocker):
             tests="some/tests/folder",
             working_dir=temp_dir,
             with_custom_airflow_packages=False,
+            with_custom_test_packages=False,
         )
 
         assert spy_writer.call_count == 1